extract the form out of the controller, and put a big comment to ensure it's clear that part is used by mobile
