+++ app/controllers/StripeController.scala
@@ -7,19 +7,17 @@ import actions.CommonActions._
 import cats.data.EitherT
 import cats.syntax.show._
 import com.gu.i18n.CountryGroup._
-import com.gu.i18n.{AUD, Currency, EUR, USD}
+import com.gu.i18n.{AUD, EUR, USD}
 import com.gu.stripe.Stripe
 import com.gu.stripe.Stripe.Charge
 import com.gu.stripe.Stripe.Serializer._
 import com.typesafe.config.Config
+import controllers.forms.ContributionRequest
 import cookies.ContribTimestampCookieAttributes
 import cookies.syntax._
 import models._
 import org.joda.time.DateTime
 import play.api.Logger
-import play.api.data.Forms._
-import play.api.data.format.Formatter
-import play.api.data.{Form, FormError}
 import play.api.libs.json._
 import play.api.mvc._
 import services.PaymentServices
@@ -30,51 +28,9 @@ import scala.concurrent.{ExecutionContext, Future}
 class StripeController(paymentServices: PaymentServices, stripeConfig: Config)(implicit ec: ExecutionContext)
   extends Controller with Redirect {
 
-  implicit val currencyFormatter = new Formatter[Currency] {
-    type Result = Either[Seq[FormError], Currency]
-
-    override def bind(key: String, data: Map[String, String]): Result =
-      data.get(key).map(_.toUpperCase).flatMap(Currency.fromString).fold[Result](Left(Seq.empty))(currency => Right(currency))
-
-    override def unbind(key: String, value: Currency): Map[String, String] =
-      Map(key -> value.identifier)
-  }
-
-  case class ContributionRequest(
-    name: String,
-    currency: Currency,
-    amount: BigDecimal,
-    email: String,
-    token: String,
-    marketing: Boolean,
-    postcode: Option[String],
-    ophanPageviewId: String,
-    ophanBrowserId: Option[String],
-    cmp: Option[String],
-    intcmp: Option[String],
-    refererPageviewId: Option[String],
-    refererUrl: Option[String]
-  )
-
-  val contributionForm: Form[ContributionRequest] = Form(
-    mapping(
-      "name" -> text,
-      "currency" -> of[Currency],
-      "amount" -> bigDecimal(10, 2),
-      "email" -> email,
-      "token" -> nonEmptyText,
-      "marketing" -> boolean,
-      "postcode" -> optional(nonEmptyText),
-      "ophanPageviewId" -> text,
-      "ophanBrowserId" -> optional(text),
-      "cmp" -> optional(text),
-      "intcmp" -> optional(text),
-      "refererPageviewId" -> optional(text),
-      "refererUrl" -> optional(text)
-    )(ContributionRequest.apply)(ContributionRequest.unapply)
-  )
-
-  def pay = (NoCacheAction andThen MobileSupportAction andThen ABTestAction).async(BodyParsers.jsonOrMultipart(contributionForm)) { implicit request =>
+  // THIS ENDPOINT IS USED BY BOTH THE FRONTEND AND THE MOBILE-APP
+  def pay = (NoCacheAction andThen MobileSupportAction andThen ABTestAction)
+    .async(BodyParsers.jsonOrMultipart(ContributionRequest.contributionForm)) { implicit request =>
 
     val form = request.body
